+++ CHANGELOG.md
@@ -8,6 +8,10 @@ All notable changes to this project will be documented in this file.
 
 - Remove unused `IexecLibOrders_v5` generated class. (#68)
 
+### Quality
+
+- Use `@SneakyThrows` lombok annotation in `EIP-712` related tests. (#69)
+
 ## [[3.2.0]](https://github.com/iExecBlockchainComputing/iexec-commons-poco/releases/tag/v3.2.0) 2023-12-19
 
 ### New Features